requirements.txt
flask==3.1.1
supabase==2.16.0
python-dotenv==1.1.1
isbnlib==3.10.14
pytest==8.4.1
setuptools==80.9.0
flask-cors==6.0.1
pyzbar==0.1.9
opencv-python==4.12.0.88
matplotlib==3.10.5
pandas==2.3.1
beautifulsoup4==4.13.4
requests==2.32.5
selenium==4.35.0
pdfplumber==0.11.7
pytesseract==0.3.13

langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.40
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langsmith==0.3.11
pypdf==5.3.0
# streamlit==1.42.2

scikit-learn==1.7.1

flask-swagger-ui==5.21.0
Flask-RESTful==0.3.10
flask-swagger==0.2.14
flasgger==0.9.7.1

# Note: Ensure that Tesseract OCR is installed on your system for pytesseract to work.